[Merge] Merge rt-tests into main (#107)
* [...] Idk

* [...] Misc changes

* [Misc] Fix compile time issues, and fix FetchData

* Breaks rendering in non-2D FB mode

* [...] Misc changes to rendering

* [Render] OMFG WE DID IT, RENDERING!

* [...] Improve GPU structs

* [AST] Track vertex attributes properly

Hey, I'm alive?? Shocker. 😛

* [Interpreter]: Return fake EDRAM Training result.

* [Merge] Fix build issues from latest changes

* [Misc] Restructure backends, fix issues with build

* [Render] Change code to handle shaders

* [Render] Fix issues, ready to merge to main

* [Render] Fix issue with shader factory

* [XCPU] Fix issue with merge for PPCInterpreter.cpp

* [VXU] I fucking hate SSE3

* [...] Fix cli-only, fix issue with UART disconnection

* [...] Fix UART Windows build, fix VXU

* [...] Fix HDD not disabling properly, improve VXU speed

* [VXU] Revert changes

* [GFX] Fix issue in CP with render ifdef for cli-only

* [Warn] Fix issue with opcode lookup

---------

+++ Xenon/Render/Abstractions/Buffer.h
@@ -21,17 +21,17 @@ enum class eBufferType : u8 {
 class Buffer {
 public:
   virtual ~Buffer() = default;
-  virtual void CreateBuffer(u32 size, const void *data, eBufferUsage usage, eBufferType type) = 0;
-  virtual void UpdateBuffer(u32 offset, u32 size, const void *data) = 0;
+  virtual void CreateBuffer(u64 size, const void *data, eBufferUsage usage, eBufferType type) = 0;
+  virtual void UpdateBuffer(u64 offset, u64 size, const void *data) = 0;
   virtual void Bind(u32 binding = 1) = 0;
   virtual void Unbind() = 0;
   virtual void DestroyBuffer() = 0;
-  virtual void SetSize(u32 size) { Size = size; }
-  virtual u32 GetSize() const { return Size; }
+  virtual void SetSize(u64 size) { Size = size; }
+  virtual u64 GetSize() const { return Size; }
   virtual void SetType(eBufferType type) { Type = type; }
   virtual eBufferType GetType() const { return Type; }
 protected:
-  u32 Size = 0;
+  u64 Size = 0;
   eBufferType Type;
 };
